% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_spm.R
\name{aggregate_aux}
\alias{aggregate_aux}
\title{Internal use only}
\usage{
aggregate_aux(
  x,
  by,
  FUN,
  ...,
  do_union = TRUE,
  simplify = TRUE,
  join = st_intersects
)
}
\arguments{
\item{x}{internal use}

\item{by}{internal use}

\item{FUN}{internal use}

\item{...}{internal use}

\item{do_union}{internal use}

\item{simplify}{internal use}

\item{join}{internal use}
}
\description{
{aggregate.sf} taken from \code{sf}.
}
\keyword{internal}
