% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{aux_gauss}
\alias{aux_gauss}
\title{Mean of a (Gaussian) covariance function (Internal use)}
\usage{
aux_gauss(dist, sigsq, phi)
}
\arguments{
\item{dist}{a numeric matrix representing the distance between spatial
entities.}

\item{sigsq}{the \eqn{\sigma^2} parameter from the Gaussian covariance
function.}

\item{phi}{the \eqn{\phi} parameter from the Gaussian covariance function,
controls the range of the spatial dependence.}
}
\value{
The mean of \code{gauss_cov(dist, sigsq, phi)}.
}
\description{
This is an auxilliary function for internal use. It helps to
numerically integrate a covariance function evaluated at a grid of points
within a polyigon and speed-up the computations.
}
\keyword{internal}
