% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{aux_spher}
\alias{aux_spher}
\title{Mean of a (Spherical) covariance function (Internal use)}
\usage{
aux_spher(dist, sigsq, phi)
}
\arguments{
\item{dist}{a numeric matrix representing the distance between spatial
entities.}

\item{sigsq}{the \eqn{\sigma^2} parameter from the Spherical covariance.
function.}

\item{phi}{the \eqn{\phi} parameter from the Spherical covariance function,
controls the range of the spatial dependence.}
}
\value{
The mean of \code{spher_cov(dist, sigsq, phi)}.
}
\description{
This is an auxiliary function for internal use. It helps to
numerically integrate a covariance function evaluated at a grid of points
within a polygon and speed-up the computations.
}
\keyword{internal}
