% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{aux_tapmat}
\alias{aux_tapmat}
\title{Mean of a (Matern - Wendland-1 tapper) covariance function (Internal use)}
\usage{
aux_tapmat(dist, sigsq, phi, nu, theta)
}
\arguments{
\item{dist}{a numeric matrix representing the distance between spatial
entities.}

\item{sigsq}{the \eqn{\sigma^2} parameter from the Matern covariance
function.}

\item{phi}{the \eqn{\phi} parameter from the Matern covariance function,}

\item{nu}{smoothness parameter}

\item{theta}{the \eqn{\theta} tapper range.}
}
\value{
The mean of \code{mat_cov(dist, sigsq, phi, nu)}.
}
\description{
This is an auxiliary function for internal use. It helps to
numerically integrate a covariance function evaluated at a grid of points
within a polygon and speed-up the computations.
}
\keyword{internal}
