% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{aux_w1}
\alias{aux_w1}
\title{Mean of a (Wendland-1) covariance function (Internal use)}
\usage{
aux_w1(dist, sigsq, phi)
}
\arguments{
\item{sigsq}{the \eqn{\sigma^2} parameter from the Matern covariance
function.}

\item{phi}{the \eqn{\phi} parameter from the Matern covariance function,}

\item{dists}{a numeric matrix representing the distance between spatial
entities.}
}
\value{
The mean of \code{mat_cov(dist, sigsq, phi, kappa)}.
}
\description{
This is an auxilliary function for internal use. It helps to
numerically integrate a covariance function evaluated at a grid of points
within a polyigon and speed-up the computations.
}
\keyword{internal}
