% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ax_fxs.R
\name{b_season}
\alias{b_season}
\title{Estimating the number of infections caused by 1 LIZ under seasonality and no dispersion effort}
\usage{
b_season(b0, b1, period, t)
}
\arguments{
\item{b0}{strength of seasonality parameter}

\item{b1}{strength of seasonality parameter}

\item{period}{periodicity of infection in weeks}

\item{t}{vector of time}
}
\value{
description
}
\description{
Introducing seasonality in the infection probability through b.
}
\examples{
b_season(-30, 0.85, 3*52, 1:10)
}
