% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ax_fxs.R
\name{calc_local_R0}
\alias{calc_local_R0}
\title{Calculating local R0}
\usage{
calc_local_R0(tau, theta, b, E)
}
\arguments{
\item{tau}{dispersion parameter}

\item{theta}{dispersion parameter}

\item{b}{number of infections caused by one LIZ}

\item{E}{Number of spores in the environment}
}
\value{
description
}
\description{
Calculating local R0
}
\examples{
calc_local_R0(10, 10, 0.001, 1)
}
