% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_survival_prob.R
\name{calc_survival_prob}
\alias{calc_survival_prob}
\title{Calculate survival probability based on vaccination rate}
\usage{
calc_survival_prob(beta_0, beta_1, vax_rate)
}
\arguments{
\item{beta_0}{vector or single value. The starting point for natural immunity. A higher value means greater natural immunity}

\item{beta_1}{vector or single value. The effect of the vaccine. Higher value means the vaccine has a greater positive effect on survival}

\item{vax_rate}{value between 0 and 1, to represent the percentage of the population that is vaccinated}
}
\value{
description
}
\description{
This function allows us to visualize and also calculate the survival probability
based on a set of parameters
}
\examples{
calc_survival_prob(beta_0 = -1, beta_1 = 10, vax_rate = seq(0, 1, 0.01))

}
