% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{comp_pexp_cov}
\alias{comp_pexp_cov}
\title{Powered Exponential covariance function for a polygons.

Computing the Powered Exponential covariance function between polygons.}
\usage{
comp_pexp_cov(cross_dists, n, n2, sigsq, phi, kappa)
}
\arguments{
\item{cross_dists}{a \code{list} such that each position contains the cross
distances between points within different polygons.}

\item{n}{an ingeger representing number of polygons (note that, this is
different than the size of the list \code{cross_dists}}

\item{n2}{usually, equal to \code{n}, except when the function is being used
to calculate the "cross" covariance between two different partitions of
the same space.}

\item{sigsq}{the \eqn{\sigma^2} parameter from the Powered Exponential
covariance function.}

\item{phi}{the \eqn{\phi} parameter from the Powered Exponential covariance
function, controls the range of the spatial dependence.}

\item{kappa}{the \eqn{\kappa \in (0, 2]} parameter representing the "power"}
}
\value{
The powered exponential covariance matrix associated with a set of
polygons.
}
\description{
Powered Exponential covariance function for a polygons.

Computing the Powered Exponential covariance function between polygons.
}
\seealso{
\code{\link{single_exp}}, \code{\link{single_matern}},
\code{\link{mat_cov}}
}
