% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_spm.R
\name{compute_mats}
\alias{compute_mats}
\alias{comp_haus_mats}
\alias{comp_grid_mats}
\title{Computing the matrices to make predictions.}
\usage{
comp_haus_mats(dists, cdists, predists, model, phi, sigsq, kappa)

comp_grid_mats(
  dists,
  cdists,
  predists,
  model,
  phi,
  sigsq,
  kappa,
  n_obs,
  n_pred
)
}
\arguments{
\item{dists}{either a \code{numeric matrix} representing the (hausdorff)
distance matrix between polygons or a \code{list} used to compute the
"grid covariances".}

\item{cdists}{either a \code{numeric matrix} representing the (hausdorff)
distance matrix between observed polygons and predicted locations or a
\code{list} used to compute the "grid cross-covariances".}

\item{predists}{a \code{numeric matrix} representing the (hausdorff)
distance matrix between the prediction locations.}

\item{model}{a \code{character} scalar indicating the family of the
covariance function to be used. The options are \code{c("matern",
    "pexp", "gaussian", "spherical")}.}

\item{phi}{a \code{numeric} value indicating the \eqn{\theta} parameter
associated with the "decay" speed of the spatial dependence.}

\item{sigsq}{a \code{numeric} value indicating the \eqn{\sigma^2} parameter
associated with the variance of a Gaussian Process.}

\item{kappa}{a \code{numeric} value indicating either the \eqn{\kappa}
paramater from the Matern covariance function (controlling the process
differentiability), or the "pexp" for the Powered Exponential family. If
the \code{model} chosen by the user is Matern and \code{kappa} is not
informed, it is automatically set to .5. On the other hand, if the user
choses the Powered Exponential family and do not inform \code{kappa},
then it is set to 1. In both cases, the covariance function becomes the
so covalled exponential covariance function.}

\item{n_obs}{number of observed polygons (used only when \code{method ==
"grid"}).}

\item{n_pred}{number of prediction locations (used only when \code{method ==
"grid"}).}
}
\value{
a \code{list} containing three covariance matrices associated with
the observed data, cross covariance between observed data and locations
on which we want to make the predictions, and the predicted locations.
}
\description{
Computes the matrices needed to make predictions.
}
