% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{crossdist}
\alias{crossdist}
\title{Pairwise distances between two matrices (Eigen version)}
\usage{
crossdist(m1, m2)
}
\arguments{
\item{m1}{A matrix where each row is a 2D coordinate.}

\item{m2}{A matrix where each row is a 2D coordinate.}
}
\description{
Computes Euclidean distance between rows of m1 and rows of m2.
}
\keyword{internal}
