% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{distmat}
\alias{distmat}
\title{Creating a symmetric distance matrix (Eigen version)}
\usage{
distmat(my_mat)
}
\arguments{
\item{my_mat}{A matrix where each row is a 2D coordinate.}
}
\description{
Computes the Euclidean distance between all pairs of rows in a
matrix.
}
\keyword{internal}
