% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{goodness_of_fit}
\alias{goodness_of_fit}
\alias{AIC.spm_fit}
\alias{BIC.spm_fit}
\title{Akaike's (and Bayesian) An Information Criterion for \code{spm_fit}
objects.}
\usage{
\method{AIC}{spm_fit}(x, ..., k = 2)

\method{BIC}{spm_fit}(x, ...)
}
\arguments{
\item{x}{a \code{spm_fit} object.}

\item{...}{optionally more fitted model objects.}

\item{k}{\code{numeric}, the \emph{penalty} per parameter to be used; the default
'k = 2' is the classical AIC. (for compatibility with \code{stats::AIC}.}
}
\value{
a \code{numeric} scalar corresponding to the goodness of fit
measure.
}
\description{
Akaike's (and Bayesian) An Information Criterion for \code{spm_fit}
objects.
}
