% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log-lik.R
\name{grid_log_lik}
\alias{grid_log_lik}
\title{Evaluate log-lik (when using a grid of points)}
\usage{
grid_log_lik(theta, .dt, dists, nap, model, kappa = NULL, apply_exp = FALSE)
}
\arguments{
\item{theta}{a \code{list} of size \eqn{2 p + \frac{p(p + 1)}{2} + 3}
containing the parameters associated with the model. (Explain why)}

\item{.dt}{a \code{list} with two positions. The first containing the
numerical data for the variable \eqn{Y}, and the second for \eqn{X}.}

\item{dists}{a \code{list} of size three. The first containing the distance
matrices associated with the regions where \eqn{Y} was measured, the
second for the distance matrices associated with \eqn{X}, and the last
containing the cross-distance matrices.}

\item{nap}{a \code{integer vector} containing either the number of pixels
within each polygon or the area associated with each polygon. The first
is used when we use a grid of points within the polygons, while the second
one is used when we compute the distance between the polygons using the
hausdorff distance (Ordered by the id variables for the polygons).}

\item{model}{a \code{character} indicating which covariance function to
use. Possible values are \code{c("matern", "pexp", "gaussian",
    "spherical")}.}

\item{kappa}{\eqn{\kappa} parameter. Not necessary if \code{mode} is
\code{"gaussian"} or \code{"spherical"}}

\item{apply_exp}{a \code{logical} indicater wheter the exponential
transformation should be applied to variance parameters. This
facilitates the optimization process.}
}
\value{
a scalar representing \code{-log.lik}.
}
\description{
Evaluate the log-likelihood for a given set of parameters
}
\details{
Internal use.
}
