% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spm.R
\name{haus_spm}
\alias{haus_spm}
\title{single \code{sf} to \code{spm} - hausdorff}
\usage{
haus_spm(sf_obj, poly_ids = NULL, var_ids = NULL)
}
\arguments{
\item{sf_obj}{a \code{sf} object s.t. its geometries are polygons.}

\item{poly_ids}{a \code{character} vector informing the name of the variable
in \code{sf_obj} that represents the polygons unique identifiers. In
case this is not informed, we assume the id of the polygons are given by
their row numbers.}

\item{var_ids}{a scalar or vector of type \code{character} indicating the
(numerical) variables that are going to be analyzed.}
}
\value{
a \code{list}.
}
\description{
Transforming a \code{sf} into a \code{spm} object (Internal
use)
}
\details{
This is a helper function for internal use. It is used within
\code{\link{single_sf_to_spm}} when \code{method = "hausdorff"} is set.
}
