% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ax_fxs.R
\name{lambda_t}
\alias{lambda_t}
\title{Calculate infection probability for time t}
\usage{
lambda_t(theta, tau, b, E)
}
\arguments{
\item{theta}{disease dispersion parameter}

\item{tau}{disease dispersion parameter}

\item{b}{number of infections caused by one LIZ}

\item{E}{Number of spores in the environment}
}
\value{
description
}
\description{
Infection probability based on Ponciano and Capistran 2011.
}
\examples{
lambda_t(10, 10, 0.001, 1)
}
