% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{liv_lsoa}
\alias{liv_lsoa}
\title{Liverpool Lower Super Output Area.}
\format{
A \code{sf} data frame with 298 rows and 6 variables:
\describe{
\item{lsoa11cd}{LSOA code}
\item{lsoa11cd}{LSOA name} \item{male}{Male population}
\item{female}{Female population}
\item{imdscore}{Index of Multiple Deprivation}
\item{area}{LMSOA area, in \eqn{km^2}}
}
}
\source{
\url{https://ij-healthgeographics.biomedcentral.com/articles/10.1186/s12942-020-00200-w}
}
\usage{
liv_lsoa
}
\description{
A dataset containing containing the LSOA's for Liverpool along
with estimates for Index of Multiple Deprivation. Data taken from
\href{https://ij-healthgeographics.biomedcentral.com/articles/10.1186/s12942-020-00200-w}{Johnson et al. 2020}
}
\details{
The data was projected to EPSG 27700 and units changed to km
}
\keyword{datasets}
