% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation_fx.R
\name{LIZ_negll}
\alias{LIZ_negll}
\title{LIZ negative log likelihood}
\usage{
LIZ_negll(pars = c(theta, tau, b0, b1), period, years, SMILE.obs)
}
\arguments{
\item{pars}{list of parameters to estimate}

\item{period}{periodicity of infection in weeks}

\item{years}{number of years}

\item{SMILE.obs}{observation data}
}
\value{
description
}
\description{
to be used with the estimation function
}
