% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log-lik.R
\name{log_lik_spm}
\alias{log_lik_spm}
\title{Evaluate log-lik}
\usage{
log_lik_spm(
  theta,
  .dt,
  dists,
  npix,
  model,
  nu = NULL,
  apply_exp = FALSE,
  type = c("full", "profile")
)
}
\arguments{
\item{theta}{a named \code{numeric} vector containing parameters (\eqn{\mu, \sigma^2, \alpha, \phi}).}

\item{.dt}{a \code{numeric} vector containing the variable \eqn{Y}.}

\item{dists}{a \code{list} of distance matrices.}

\item{npix}{a \code{integer vector} containing the number of pixels within each polygon.}

\item{model}{a \code{character} indicating the covariance function family.}

\item{nu}{\eqn{\nu} parameter for the Matern or Powered Exponential covariance functions.}

\item{apply_exp}{a \code{logical} indicating whether to exponentiate non-negative parameters.}

\item{type}{a \code{character} specifying the likelihood type: \code{"full"} or \code{"profile"}.}
}
\value{
a scalar representing \code{-log.lik}.
}
\description{
Unified function for log-likelihood evaluation
}
\details{
Internal use.
}
\keyword{internal}
