% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mat_cov}
\alias{mat_cov}
\alias{comp_mat_cov}
\alias{pexp_cov}
\alias{comp_pexp_cov}
\title{Matern covariance function for a given distance matrix.}
\usage{
mat_cov(dists, sigsq, phi, nu)

comp_mat_cov(cross_dists, n, n2, sigsq, phi, nu)

pexp_cov(dists, sigsq, phi, nu)

comp_pexp_cov(cross_dists, n, n2, sigsq, phi, nu)
}
\arguments{
\item{dists}{a numeric matrix representing the distance between spatial
entities.}

\item{sigsq}{the \eqn{\sigma^2} parameter from the Matern covariance
function.}

\item{phi}{the \eqn{\phi} parameter from the Matern covariance function,
controls the range of the spatial dependence.}

\item{nu}{the \eqn{\nu} parameter from the Matern covariance function,
controls the differentiability of the process.}

\item{cross_dists}{a \code{list} such that each position contains the cross
distances between points within different polygons.}
}
\value{
The matern covariance function (for a stationary and isotropic
process) associated with the provided distances (\code{dists}) and the
given set of parameters.

The matern covariance matrix associated with a set of polygons.
}
\description{
Computing the Matern covariance function for a matrix of
distances.

Computing the Matern covariance function between polygons.
}
\seealso{
\code{\link{single_exp}}, \code{\link{single_matern}}

\code{\link{single_exp}}, \code{\link{single_matern}},
\code{\link{mat_cov}}
}
\keyword{internal}
