% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sai.R
\name{morans_i}
\alias{morans_i}
\title{Calculates the (global) Moran's I}
\usage{
morans_i(sf_dt, variable)
}
\arguments{
\item{sf_dt}{a \code{sf} (with POLYGON \code{geometry}) dataset.}

\item{variable}{a \code{character} representing one of the variables from
\code{sf_dt}.}
}
\value{
a \code{numeric} scalar.
}
\description{
Calculates the (global) Moran's I
}
