% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mult_dists}
\alias{mult_dists}
\title{Cross-distances for two lists of matrices (Eigen version)}
\usage{
mult_dists(mat_list1, mat_list2, return_single)
}
\description{
Cross-distances for two lists of matrices (Eigen version)
}
\keyword{internal}
