% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{nl_ct}
\alias{nl_ct}
\title{Nova Lima census tracts}
\format{
A \code{sf} data frame with 113 rows and 14 variables:
\describe{
\item{cd_setor}{unique identifier}
\item{hh_density}{average household density}
\item{var_hhd}{variance of the household density}
\item{avg_income}{average income per household}
\item{var_income}{variance of the income per household}
\item{pop}{population in the census tract}
\item{avg_age}{average age of the inhabitants in the census tract}
\item{var_age}{variance of the variable age in the census tract}
\item{prop_women}{proportion of women}
\item{prop_elder}{proportion of people with 65 years of age or older}
\item{illit_rate}{illiteracy rate}
\item{prop_white}{proportion of self-declared white people}
\item{prop_black}{proportion of self-declared black people}
\item{prop_native}{proportion of self-declared native people}
}
}
\usage{
nl_ct
}
\description{
A dataset containing containing the census tracts for the city
of Nova Lima in Minas Gerais - Brazil.
}
\details{
The data is project using the SIRGAS 2000.
}
\keyword{datasets}
