% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{nyc_comd}
\alias{nyc_comd}
\title{New York City community districts spatial geometries}
\format{
A \code{sf} data frame with 71 rows and 3 variables:
\describe{
\item{boro_cd}{unique identifier}
\item{shape_area}{Shape Area}
\item{shape_length}{Shape Length}
\item{est}{median income estimated using areal interpolation}
\item{se_est}{standard error associated with the estimates}
}
}
\usage{
nyc_comd
}
\description{
A dataset containing containing the CD's for New York City.
}
\details{
The data is project using EPSG 4326.
}
\keyword{datasets}
