% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{nyc_surv}
\alias{nyc_surv}
\title{New York City survey data.}
\format{
A \code{sf} data frame with 2128 rows and 5 variables:
\describe{
\item{GEOID}{unique identifier}
\item{NAME}{census tract name}
\item{variable}{variable estimated}
\item{estimate}{median income estimate}
\item{moe}{median income estimate margin of error}
}
}
\usage{
nyc_surv
}
\description{
A dataset containing containing the census tracts for New York
City along with estimates for median income and a margin of error for
this estimates.
}
\details{
The data is project using EPSG 4326.
}
\keyword{datasets}
