% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ax_fxs.R
\name{rho_n}
\alias{rho_n}
\title{Calculate average reproduction rate at time t}
\usage{
rho_n(N, K, rho = 0.41)
}
\arguments{
\item{N}{population size at time t}

\item{K}{carrying capacity}

\item{rho}{average reproduction rate}
}
\value{
description
}
\description{
Calculate average reproduction rate at time t
}
\examples{
rho_n(2000, 5000)
}
