% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sai.R
\name{SAI}
\alias{SAI}
\alias{sai}
\alias{sai_var}
\title{Simple Areal Interpolation}
\usage{
sai(source, target, vars)

sai_var(source, target, vars, vars_var, var_method = "CS")
}
\arguments{
\item{source}{a \code{sf} object - source spatial data.}

\item{target}{a \code{sf} object - target spatial data.}

\item{vars}{a \code{character} representing the variables (observed at the
source) to be estimated at the target data.}

\item{vars_var}{a scalar of type \code{character} representing the name of
the variable in the source dataset that stores the variances of the
variable to be estimated at the target data.}

\item{var_method}{a \code{character} representing the method to approximate
the variance of the SAI estimates. Possible values are "CS"
(Cauchy-Schwartz) or "MI" (Moran's I).}
}
\value{
the target (of type \code{sf}) with estimates of the variables
observed at the source data.
}
\description{
Simple Areal Interpolation
}
