% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{single_dists}
\alias{single_dists}
\title{Pairwise distances for a list of matrices (Eigen version)}
\usage{
single_dists(mat_list)
}
\description{
Pairwise distances for a list of matrices (Eigen version)
}
\keyword{internal}
