% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{single_dists_tr}
\alias{single_dists_tr}
\title{Truncated pairwise distances for a list of matrices (Eigen version)}
\usage{
single_dists_tr(mat_list, tr_vec, tr_inp)
}
\description{
Truncated pairwise distances for a list of matrices (Eigen version)
}
\keyword{internal}
