% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{single_exp}
\alias{single_exp}
\title{Exponential covariance function (scalar)

Computing the Exponential covariance function for a single distance measure.}
\usage{
single_exp(d, sigsq, phi)
}
\arguments{
\item{d}{a scalar representing the distance on which it is desired to
evaluate the covariance function.}

\item{sigsq}{the \eqn{\sigma^2} parameter from the Exponential covariance
function.}

\item{phi}{the \eqn{\phi} parameter from the Exponential covariance function,
controls the range of the spatial dependence.}
}
\value{
a scalar representing the (exponential) covariance between two
observations \code{d} apart of each other.
}
\description{
Exponential covariance function (scalar)

Computing the Exponential covariance function for a single distance measure.
}
\seealso{
\code{\link{single_pexp}}, \code{\link{single_matern}},
\code{\link{single_matern3}}, \code{\link{single_matern5}},
\code{\link{mat_cov}}
}
