% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{single_gauss}
\alias{single_gauss}
\title{Gaussian covariance function (scalar)

Computing the Gaussian covariance function for a single distance measure.}
\usage{
single_gauss(d, sigsq, phi)
}
\arguments{
\item{d}{a scalar representing the distance on which it is desired to
evaluate the covariance function.}

\item{sigsq}{the \eqn{\sigma^2} parameter from the Gaussian covariance
function.}

\item{phi}{the \eqn{\phi} parameter from the Gaussian covariance function,
controls the range of the spatial dependence.}
}
\value{
a scalar representing the (gaussian) covariance between two
observations \code{d} apart of each other.
}
\description{
Gaussian covariance function (scalar)

Computing the Gaussian covariance function for a single distance measure.
}
\seealso{
\code{\link{single_exp}}, \code{\link{single_matern}},
\code{\link{single_matern3}}, \code{\link{single_matern5}},
\code{\link{mat_cov}}
}
