% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{single_gw0}
\alias{single_gw0}
\title{Matern Generalized Wendland (GW) covariance function with kappa = 0
(scalar - generic)}
\usage{
single_gw0(d, sigsq, phi, mu)
}
\arguments{
\item{d}{a scalar representing the distance on which it is desired to
evaluate the covariance function.}

\item{sigsq}{the \eqn{\sigma^2} parameter from the Matern covariance
function.}

\item{phi}{the \eqn{\phi} parameter from the Matern covariance function,
controls the range of the spatial dependence.}

\item{mu}{a parameter that controls the smoothness of the covariance
function. Note that, \eqn{\mu \geq 1}.}
}
\value{
a scalar representing the GW covariance between two
observations \code{d} apart of each other.
}
\description{
adapted from Bevilacqua et al. 2019.
}
\keyword{internal}
