% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{single_matern}
\alias{single_matern}
\title{Matern covariance function (scalar - generic)

Computing the Matern covariance function for a single distance measure,
addapted from \code{\link[geoR]{matern}}.}
\usage{
single_matern(d, sigsq, phi, kappa)
}
\arguments{
\item{d}{a scalar representing the distance on which it is desired to
evaluate the covariance function.}

\item{sigsq}{the \eqn{\sigma^2} parameter from the Matern covariance
function.}

\item{phi}{the \eqn{\phi} parameter from the Matern covariance function,
controls the range of the spatial dependence.}

\item{kappa}{the \eqn{\kappa} parameter from the Matern covariance function,
controls the differentiability of the process.}
}
\value{
a scalar representing the (matern) covariance between two
observations \code{d} apart of each other.
}
\description{
Matern covariance function (scalar - generic)

Computing the Matern covariance function for a single distance measure,
addapted from \code{\link[geoR]{matern}}.
}
\seealso{
\code{\link{single_matern}}, \code{\link{single_matern5}}
\code{\link{single_exp}}, \code{\link{mat_cov}}
}
