% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{single_spher}
\alias{single_spher}
\title{Spherical covariance function (scalar)}
\usage{
single_spher(d, sigsq, phi)
}
\arguments{
\item{d}{a scalar representing the distance on which it is desired to
evaluate the covariance function.}

\item{sigsq}{the \eqn{\sigma^2} parameter from the Spherical covariance.
function.}

\item{phi}{the \eqn{\phi} parameter from the Spherical covariance function,
controls the range of the spatial dependence.}
}
\value{
a scalar representing the (gaussian) covariance between two
observations \code{d} apart of each other.
}
\description{
Computing the Spherical covariance function for a scalar
distance.
}
\seealso{
\code{\link{single_exp}}, \code{\link{single_matern}},
\code{\link{single_matern3}}, \code{\link{single_matern5}},
\code{\link{mat_cov}}
}
\keyword{internal}
