% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{single_w1}
\alias{single_w1}
\title{Matern Wendland-1 covariance unction(scalar - generic)}
\usage{
single_w1(d, sigsq, phi)
}
\arguments{
\item{d}{a scalar representing the distance on which it is desired to
evaluate the covariance function.}

\item{sigsq}{the \eqn{\sigma^2} parameter from the Matern covariance
function.}

\item{phi}{the \eqn{\phi} parameter from the Matern covariance function,
controls the range of the spatial dependence.}
}
\value{
a scalar representing the (wendland) covariance between two
observations \code{d} apart of each other.
}
\description{
adapted from Furrer et al. 2006.
}
\keyword{internal}
