% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMILE-package.R
\docType{package}
\name{SMILE-package}
\alias{SMILE}
\alias{SMILE-package}
\title{SMILE: Compartmental model for anthrax.}
\description{
Code from our papers to model anthrax dynamics using compartmental models.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ufseerlab.github.io/SMILE/}
}

}
\author{
\strong{Maintainer}: javi rudolph \email{javiera.av@ufl.edu}

}
\keyword{internal}
