% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smile_fx.R
\name{smile_fx}
\alias{smile_fx}
\title{SMILE function}
\usage{
smile_fx(
  b0,
  b1,
  period,
  theta,
  tau,
  years,
  alpha = 1/52,
  zeta_novax = 0.88,
  gamma = 0.9868,
  sigmaa = 1,
  psi = 1,
  N1 = 5000,
  K = 5000,
  vax = NULL,
  beta_0 = NULL,
  beta_1 = NULL,
  b_fixed = NULL,
  age_struc = TRUE,
  stochastic = FALSE,
  LIZ_init = 1,
  rho_pop = NULL,
  output_df = FALSE
)
}
\arguments{
\item{output_df}{}
}
\description{
This function encompasses the original functions and assumptions. It outputs a time series for each of the compartments given a set of parameters.
}
