% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation_fx.R
\name{SMILE_param_estim}
\alias{SMILE_param_estim}
\title{Parameter estimation from data}
\usage{
SMILE_param_estim(b0, b1, theta, tau, period, SMILE.obs, method = "BFGS")
}
\arguments{
\item{b0}{starting value}

\item{b1}{starting value}

\item{theta}{starting value}

\item{tau}{starting value}

\item{period}{starting value}

\item{SMILE.obs}{data}

\item{method}{set to BFGS}
}
\value{
description
}
\description{
Parameter estimation from data
}
