% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{spher_cov}
\alias{spher_cov}
\title{Computing the Spherical covariance function for a single distance measure.}
\usage{
spher_cov(dists, sigsq, phi)
}
\arguments{
\item{dists}{a numeric matrix representing the distance between spatial
entities.}

\item{sigsq}{the \eqn{\sigma^2} parameter from the Spherical covariance.
function.}

\item{phi}{the \eqn{\phi} parameter from the Spherical covariance function,
controls the range of the spatial dependence.
'}
}
\value{
The Spherical covariance function (for a stationary and
isotropic process) associated with the provided distances (\code{dists})
and the given set of parameters.
}
\description{
Computing the Spherical covariance function for a single distance measure.
}
\seealso{
\code{\link{single_exp}}, \code{\link{single_matern}}
}
