% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spmismo-package.R
\docType{package}
\name{spmismo}
\alias{spmismo}
\title{spmismo: A package for modelling spatially misaligned data}
\description{
ORGANIZE THE FUNCTIOSN IN SECTIONS
The foo package provides three categories of important functions:
foo, bar and baz.
}
\section{Foo functions}{

The foo functions ...
}

