% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{st_remove_holes}
\alias{st_remove_holes}
\title{Remove holes from a \code{sfc} POLYGON}
\usage{
st_remove_holes(x)
}
\arguments{
\item{x}{a \code{sf} or \code{sfc} polygon.}
}
\value{
a \code{sf} or \code{sfc} polygon.
}
\description{
internal use. Taken from
\url{https://cran.r-project.org/web/packages/nngeo/}
}
\keyword{internal}
