% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{summary_sspm_fit}
\alias{summary_sspm_fit}
\title{summarizing \code{spm_fit}}
\usage{
summary_sspm_fit(x, sig = 0.05)
}
\arguments{
\item{x}{either a \code{sspm_fit} or a \code{mspm_fit} object}

\item{sig}{signigicance level}
}
\description{
summarizing \code{spm_fit}
}
