% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vdl.R
\name{transform.sf}
\alias{transform.sf}
\title{Transform method for \code{sf} objects}
\usage{
\method{transform}{sf}(`_data`, ...)
}
\arguments{
\item{_data}{a \code{sf} object}

\item{...}{additional options}
}
\value{
an \code{sf} object.
}
\description{
Internal usage.
}
\keyword{internal}
