% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vdl.R
\name{vdl}
\alias{vdl}
\title{Voronoi Data Linkage}
\usage{
vdl(
  coords_sf,
  areal_sf,
  id_coords = NULL,
  list_vars = NULL,
  show_warning = TRUE,
  buff
)
}
\arguments{
\item{coords_sf}{\code{sf} object representing point coordinates}

\item{areal_sf}{\code{sf} object representing the areal data (e.g. spatial polygons)}

\item{id_coords}{\code{character} indicating the \emph{id} variable for \code{coords_sf}. If
not informed, the \emph{row number} is assumed to by the id.}

\item{list_vars}{a named \code{list} representing the variables from \code{areal_sf}
that need to be estimated in \code{coords_sf}. The \code{list} position names represent
the populational variables, while the contend within each position represent
the average variavles.}

\item{show_warning}{\code{logical}, indicating wheter a warning for the crs
transformation should be printed or not. (default is \code{TRUE})}

\item{buff}{scalar \code{numeric}. Mostly for internal use.}
}
\value{
a \code{sf} object, contaning the \code{id_coords} variable and the
\code{list_vars} for the \code{coords_sf} spatial data set.
}
\description{
Reminder, have to create an example.
}
