% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vdl.R
\name{vdl_var}
\alias{vdl_var}
\title{Voronoi Data Linkage - Single variable and variance}
\usage{
vdl_var(coords_sf, areal_sf, id_coords = NULL, res_var, variance, buff)
}
\arguments{
\item{coords_sf}{\code{sf} object representing point coordinates}

\item{areal_sf}{\code{sf} object representing the areal data (e.g. spatial
polygons)}

\item{id_coords}{\code{character} indicating the \emph{id} variable for
\code{coords_sf}. If not informed, the \emph{row number} is assumed to by the id.}

\item{res_var}{a \code{character} - the name of the variable in the
\code{areal_sf} to be estimated in the \code{coords_sf}.}

\item{variance}{a \code{character} - the name of the variable varinace in
the \code{areal_sf} to be estimated in the \code{coords_sf}.}

\item{buff}{scalar \code{numeric}. Mostly for internal use.}
}
\value{
a \code{sf} object, contaning the \code{id_coords} variable and the
\code{list_vars} for the \code{coords_sf} spatial data set.
}
\description{
Reminder, have to create an example.
}
