% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{viz_dt}
\alias{viz_dt}
\title{Auxilliary dataset - Internal use}
\format{
A \code{sf} data frame with 2199 rows and 2 variables:
\describe{
\item{source}{dataset identifier}
\item{estimate}{estimated median income in NYC}
}
}
\usage{
viz_dt
}
\description{
Used to decrease compilation time.
}
\keyword{internal}
