% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vdl.R
\name{vor_build}
\alias{vor_build}
\title{Voronoi Tessellation inside a polygon}
\usage{
vor_build(points_sf, poly_sf)
}
\arguments{
\item{points_sf}{\code{sf data frame} containing the points' coordinates}

\item{poly_sf}{a \code{sf} polygon}
}
\value{
a \code{sf} object containing the polygons associated with the
voronoi tessellation of \code{points_sf} the polygon \code{poly_sf}
}
\description{
voronoi tessellation of a given a set of points inside a
polygon.  This is an internal use function.
}
\keyword{internal}
