% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{w1_cov}
\alias{w1_cov}
\title{Wendland-1 covariance function for a given distance matrix.}
\usage{
w1_cov(dists, sigsq, phi)
}
\arguments{
\item{dists}{a numeric matrix representing the distance between spatial
entities.}

\item{sigsq}{the \eqn{\sigma^2} parameter from the Matern covariance
function.}

\item{phi}{the \eqn{\phi} parameter from the Matern covariance function,'}
}
\value{
The matern covariance function (for a stationary and isotropic
process) associated with the provided distances (\code{dists}) and the
given set of parameters.
}
\description{
Computing the Matern covariance function for a matrix of
distances.
}
\keyword{internal}
