% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sai.R
\name{AI}
\alias{AI}
\alias{ai}
\alias{ai_var}
\title{Areal Interpolation}
\usage{
ai(source, target, vars)

ai_var(source, target, vars, vars_var, sc_vars = FALSE, var_method = "CS")
}
\arguments{
\item{source}{a \code{sf} object - source spatial data.}

\item{target}{a \code{sf} object - target spatial data.}

\item{vars}{a \code{character} representing the variables (observed at the
source) to be estimated at the target data.}

\item{vars_var}{a scalar of type \code{character} representing the name of
the variable in the source dataset that stores the variances of the
variable to be estimated at the target data.}

\item{sc_vars}{boolean indicating whether \code{vars} should be scaled by its
observed variance (if available).}

\item{var_method}{a \code{character} representing the method to approximate
the variance of the AI estimates. Possible values are "CS"
(Cauchy-Schwartz) or "MI" (Moran's I).}
}
\value{
the target (of type \code{sf}) with estimates of the variables
observed at the source data.
}
\description{
Areal Interpolation
}
