% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/dists-aux.R
\name{aux_mat}
\alias{aux_mat}
\alias{eucl_aux}
\alias{single_dists}
\alias{mult_dists}
\alias{pred_cdist}
\alias{get_grid_list}
\alias{dist_from_grids}
\alias{mult_dist_from_grids}
\title{Internal use only}
\usage{
eucl_aux(x, y)

single_dists(mat_list)

mult_dists(mat_list1, mat_list2, return_single)

pred_cdist(mat_list, pred_mat)

get_grid_list(x_to_list, by)

dist_from_grids(y_grid, by)

mult_dist_from_grids(y_grid, x_grid, by)
}
\arguments{
\item{x}{internal use}

\item{y}{internal use}

\item{mat_list}{internal use}

\item{mat_list1}{internal use}

\item{mat_list2}{internal use}

\item{return_single}{internal use}

\item{pred_mat}{internal use}

\item{x_to_list}{internal use}

\item{by}{internal use}

\item{y_grid}{internal use}

\item{x_grid}{internal use}
}
\description{
Internal use only
}
