% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{crossdist}
\alias{crossdist}
\title{Pairwise distances between matrices}
\usage{
crossdist(x, y)
}
\arguments{
\item{x}{a matrix representing a grid of points within a polygon.}

\item{y}{a matrix representing a grid of points within a polygon.}
}
\description{
Internal use.
}
