% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{crossdist}
\alias{crossdist}
\title{Pairwise distances between matrices}
\usage{
crossdist(m1, m2)
}
\arguments{
\item{m1}{a matrix representing a grid of points within a polygon.}

\item{m2}{a matrix representing a grid of points within a polygon.}
}
\description{
Internal use.
}
