% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{distmat}
\alias{distmat}
\title{Creating a distance matrix}
\usage{
distmat(my_mat)
}
\arguments{
\item{my_mat}{a matrix representing a grid of points.}
}
\description{
Internal use. For now it only supports euclidean distance. (May
import parallelDist in the future).
}
\keyword{internal}
