% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{est_mle}
\alias{est_mle}
\title{MLEs for fixed V.}
\usage{
est_mle(y, X, Vinv)
}
\arguments{
\item{y}{response variable.}

\item{X}{design matrix.}

\item{Vinv}{inverse of \eqn{V}}
}
\description{
MLEs for fixed V.
}
\details{
internal use.
}
