% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_phi}
\alias{find_phi}
\title{Find phi parameter for the Exponential spatial auto-correlation function}
\usage{
find_phi(
  d,
  nu,
  kappa,
  mu2,
  family = "matern",
  range = c(1e-04, 1000),
  cut = 0.05
)
}
\arguments{
\item{d}{maximun distance for spatial dependence equal to \code{cut}.}

\item{nu}{smoothness parameter associated with the Matern cov. function.}

\item{kappa}{par GW}

\item{mu2}{smoothness GW}

\item{family}{covariance function family}

\item{range}{Minimum and maximum distance to be considered.}

\item{cut}{Spatial correlation at a distance \code{d}.}
}
\value{
\code{real number}
}
\description{
Function designed to find the phi paramter such that the correlation
between points wihtin a given distance \code{d} is at most a given
value.
}
