% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find-phi.R
\name{find_phi}
\alias{find_phi}
\title{Find phi parameter for the Exponential spatial auto-correlation function}
\usage{
find_phi(d, kappa = 0.5, range = c(1e-04, 1), cut = 0.05)
}
\arguments{
\item{d}{maximun distance for spatial dependence}

\item{kappa}{parameter}

\item{range}{? Marcos help needed here}

\item{cut}{? Marcos help needed here}
}
\value{
\code{real number}
}
\description{
Function designed to find the phi paramter such that the correlation
between points wihtin a given distance \code{d} is at most a given
value.
}
