% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit_spm}
\alias{fit_spm}
\alias{fit_spm.sspm}
\alias{fit_spm.mspm}
\title{Fit a model adapted from
\href{https://ij-healthgeographics.biomedcentral.com/articles/10.1186/s12942-020-00200-w}{Johnson
et al. 2020}.}
\usage{
fit_spm(x, ...)

\method{fit_spm}{sspm}(
  x,
  model,
  theta_st,
  kappa = NULL,
  apply_exp = FALSE,
  opt_method = "Nelder-Mead",
  control_opt = list(),
  ...
)

\method{fit_spm}{mspm}(
  x,
  model,
  theta_st,
  kappa = NULL,
  apply_exp = FALSE,
  opt_method = "Nelder-Mead",
  control_opt = list(),
  ...
)
}
\arguments{
\item{x}{an object of type \code{sspm} or \code{mspm}. Note that, the
dimension of \code{theta_st} depends on the 2 factors. 1) the number of
variables being analyzed, and 2) if the input is a \code{sspm} or a
\code{mspm} object.}

\item{...}{additionnal parameters, either passed to \code{optim}.}

\item{model}{a \code{character} scalar indicating the family of the
covariance function to be used. The options are \code{c("matern",
    "pexp", "gaussian", "spherical")}.}

\item{theta_st}{a \code{numeric} (named) vector containing the initial
parameters.}

\item{kappa}{a \code{numeric} value indicating either the \eqn{\kappa}
paramater from the Matern covariance function (controlling the process
differentiability), or the "power" for the Powered Exponential
family. If the \code{model} chosen by the user is Matern and
\code{kappa} is not informed, it is automatically set to .5. On the
other hand, if the user choses the Powered Exponential family and do not
inform \code{kappa}, then it is set to 1. In both cases, the covariance
function becomes the so covalled exponential covariance function.}

\item{apply_exp}{a \code{logical} scalar indicating wheter the parameters
that cannot assume negative values should be exponentiate or not.}

\item{opt_method}{a \code{character} scalar indicating the optimization
algorithm to be used. For details, see \code{\link[stats]{optim}}.}

\item{control_opt}{a named \code{list} containing the control arguments for
the optimization algorithm to be used. For details, see
\code{\link[stats]{optim}}.}
}
\value{
a \code{sspm_fit} or a \code{mspm_fit} object.
}
\description{
Fit a model adapted from
\href{https://ij-healthgeographics.biomedcentral.com/articles/10.1186/s12942-020-00200-w}{Johnson
et al. 2020}.

Fitting the model based on
}
\details{
This function uses the \code{\link[stats]{optim}} optimization
algorithms to find the Maximum Likelihood estimators, and their standard
errors, from a model adapted from
\href{https://ij-healthgeographics.biomedcentral.com/articles/10.1186/s12942-020-00200-w}{Johnson
et al. 2020}. The function allows the user to input the control
parameters from the \code{\link[stats]{optim}} function through the
argument \code{control_opt}, which is a named list. Additionally, the
user can input lower and upper boundaries for the optimization problem,
as well as the preferred optimization algorithm (as long as it is
available at \code{\link[stats]{optim}}). The preferred algorithm is
selected by the argument \code{opt_method}. In addition to the control
of the optimization, the user can select a covariance function among the
following: Matern, Exponential, Powered Exponential, Gaussian, and
Spherical. The parameter \code{apply_exp} is a \code{logical} scalar
such that, if set to \code{TRUE}, the \eqn{\exp} function is applied to
the nonnegative parameters, allowing the optimization algorithm to
search for all the paraters over the real numbers.
}
