% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_spm.R
\name{predict_spm}
\alias{predict_spm}
\alias{predict_spm.sspm_fit}
\alias{predict_spm.mspm_fit}
\alias{predict_spm.sf}
\alias{predict_spm.matrix}
\title{Prediction over the same or a different set of regions (or points).}
\usage{
predict_spm(x, ...)

\method{predict_spm}{sspm_fit}(x, .aggregate = TRUE, ...)

\method{predict_spm}{mspm_fit}(x, ...)

\method{predict_spm}{sf}(x, spm_obj, .aggregate = TRUE, n_pts, type, ...)

\method{predict_spm}{matrix}(x, ...)
}
\arguments{
\item{x}{a \code{sf} object such that its geometris are either points or
polygons.}

\item{...}{additional parameters}

\item{.aggregate}{\code{logical}. Should the predictions be aggregated? In
case the input is only a "fit" object, the aggregation is made over the
polygons on which the original data was observed. In case the input
\code{x} is composed by \code{sf POLYGONS}, the aggregation is made over
this new partition of the study region.}

\item{spm_obj}{an object of either class \code{sspm_fit} or \code{mspm_fit}}

\item{n_pts}{a \code{numeric} scalar standing for number of points to form a
grid over the whole region to make the predictions}

\item{type}{\code{character} type of grid to be generated. See
\code{\link[sf]{st_sample}}.}
}
\value{
an object of class \code{spm_pred}
}
\description{
Realizes predictions that can be useful when researchers are
interested in predict a variable observed in one political division of a
city (or state) on another division of the same region.
}
