% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log-lik.R
\name{singl_ll_nn_hess}
\alias{singl_ll_nn_hess}
\title{Backward compatibility wrapper for full log-likelihood (Hessian)}
\usage{
singl_ll_nn_hess(theta, .dt, dists, npix, model, nu = NULL, apply_exp = FALSE)
}
\description{
Backward compatibility wrapper for full log-likelihood (Hessian)
}
\keyword{internal}
