% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log-lik.R
\name{singl_log_lik}
\alias{singl_log_lik}
\title{Backward compatibility wrapper for full log-likelihood}
\usage{
singl_log_lik(theta, .dt, dists, npix, model, nu = NULL, apply_exp = FALSE)
}
\description{
Backward compatibility wrapper for full log-likelihood
}
\keyword{internal}
