% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log-lik.R
\name{singl_log_lik_nn}
\alias{singl_log_lik_nn}
\title{Backward compatibility wrapper for profile log-likelihood (no nugget)}
\usage{
singl_log_lik_nn(theta, .dt, dists, npix, model, nu = NULL, apply_exp = FALSE)
}
\description{
Backward compatibility wrapper for profile log-likelihood (no nugget)
}
\keyword{internal}
