% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smile-package.R
\docType{package}
\name{smile}
\alias{smile}
\title{smile: Spatial MIsaLignment Estimation}
\description{
ORGANIZE THE FUNCTIOSN IN SECTIONS
The foo package provides three categories of important functions:
foo, bar and baz.
}
\section{Foo functions}{

The foo functions ...
}

