% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{summary_spm_fit}
\alias{summary_spm_fit}
\title{summarizing \code{spm_fit}}
\usage{
summary_spm_fit(x, sig = 0.05)
}
\arguments{
\item{x}{a \code{spm_fit} object}

\item{sig}{signigicance level}
}
\description{
summarizing \code{spm_fit}
}
