% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vdl.R
\name{vdl}
\alias{vdl}
\title{Voronoi Data Linkage}
\usage{
vdl(coords_sf, areal_sf, vars, buff)
}
\arguments{
\item{coords_sf}{\code{sf} POINT target dataset.}

\item{areal_sf}{\code{sf} POLYGON source dataset.}

\item{vars}{a \code{character} representing the variables (observed at the
source - polygon) to be estimated at the target data.}

\item{buff}{scalar \code{numeric}. Mostly for internal use.}
}
\value{
a \code{sf} object for the \code{coords_sf} spatial data set.
}
\description{
Reminder, have to create an example. This will be exported
after we submit the paper for publication.
}
