% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vdl.R
\name{vdl_var}
\alias{vdl_var}
\title{Voronoi Data Linkage - Single variable and variance}
\usage{
vdl_var(coords_sf, areal_sf, res_var, variance, var_method = "CS", buff)
}
\arguments{
\item{coords_sf}{\code{sf} POINT target dataset.}

\item{areal_sf}{\code{sf} POLYGON source dataset.}

\item{res_var}{a \code{character} - the name of the variable in the
\code{areal_sf} to be estimated in the \code{coords_sf}.}

\item{variance}{a \code{character} - the name of the variable varinace in
the \code{areal_sf} to be estimated in the \code{coords_sf}.}

\item{var_method}{a \code{character} representing the method to approximate
the variance of the SAI estimates. Possible values are "CS"
(Cauchy-Schwartz) or "MI" (Moran's I).}

\item{buff}{scalar \code{numeric}. Mostly for internal use.}
}
\value{
a \code{sf} object, contaning the \code{id_coords} variable and the
\code{list_vars} for the \code{coords_sf} spatial data set.
}
\description{
Reminder, have to create an example.
}
